[Refactor] 컴포넌트 분리, chatMode 리코일로 변경
검색 헤더와 초대 드롭다운에서 함께 사용되는 SearchInput을 분리하였습니다. 자주 사용되는 chatMode를 리코일로 변경하였습니다.
