+++ frontend/src/components/Chat/ChatContent/index.tsx
@@ -3,11 +3,10 @@ import { useRecoilValue, useSetRecoilState, useRecoilState } from 'recoil';
 
 import { createChatRoom } from '@apis/chat';
 import { SocketContext } from '@utils/socketContext';
-import { ChatModeType } from '@src/types/chat';
 import { UserIdType } from '@src/types/team';
 import userState from '@stores/user';
 import { teamUsersSelector } from '@stores/team';
-import { chatRoomsTrigger, currentChatRoomState, messageListState } from '@stores/chat';
+import { chatModeState, chatRoomsTrigger, currentChatRoomState, messageListState } from '@stores/chat';
 
 import { FaTelegramPlane } from 'react-icons/fa';
 import Message from './Message';
@@ -15,28 +14,21 @@ import { Container, MessagesContainer, NoticeContainer, InputContainer } from '.
 
 interface Props {
 	teamId: number;
-	chatMode: ChatModeType;
 	inviteUsers: UserIdType[];
 	messagesEndRef: React.RefObject<HTMLDivElement>;
-	setChatModeToChat: () => void;
 	initInviteUser: () => void;
 }
 
-const ChatContent: React.FC<Props> = ({
-	teamId,
-	chatMode,
-	inviteUsers,
-	messagesEndRef,
-	setChatModeToChat,
-	initInviteUser,
-}) => {
+const ChatContent: React.FC<Props> = ({ teamId, inviteUsers, messagesEndRef, initInviteUser }) => {
 	const inputRef = useRef<HTMLInputElement>(null);
 	const socketRef = useContext(SocketContext);
+
 	const myInfo = useRecoilValue(userState);
 	const teamUsers = useRecoilValue(teamUsersSelector(teamId));
-	const [currentChatRoom, setCurrentChatRoom] = useRecoilState(currentChatRoomState);
 	const setChatRoomsTrigger = useSetRecoilState(chatRoomsTrigger);
 	const messageList = useRecoilValue(messageListState);
+	const [chatMode, setChatMode] = useRecoilState(chatModeState);
+	const [currentChatRoom, setCurrentChatRoom] = useRecoilState(currentChatRoomState);
 
 	const handleNewChatRoomCreate = async () => {
 		if (!socketRef.current) return;
@@ -59,12 +51,12 @@ const ChatContent: React.FC<Props> = ({
 		socketRef.current.emit('invite users', { chatRoomId: newChatRoomInfo.chatRoomId, userList: inviteUsers, teamId });
 		inputRef.current.value = '';
 		initInviteUser();
-		setChatModeToChat();
+		setChatMode({ chatMode: 'chat' });
 	};
 
 	const handleEnterCheck = (e: React.KeyboardEvent) => {
 		if (e.key !== 'Enter') return;
-		if (chatMode === 'create') {
+		if (chatMode.chatMode === 'create') {
 			handleNewChatRoomCreate();
 			return;
 		}
@@ -84,7 +76,7 @@ const ChatContent: React.FC<Props> = ({
 
 	return (
 		<Container>
-			{chatMode === 'chat' ? (
+			{chatMode.chatMode === 'chat' ? (
 				<MessagesContainer>
 					{messageList.map((message) => (
 						<Message key={message.messageId} message={message} teamId={teamId} />
@@ -99,7 +91,7 @@ const ChatContent: React.FC<Props> = ({
 			)}
 			<InputContainer>
 				<input type='text' placeholder='새 메시지 입력' ref={inputRef} onKeyPress={handleEnterCheck} />
-				<FaTelegramPlane onClick={chatMode === 'create' ? handleNewChatRoomCreate : handleSendMessage} />
+				<FaTelegramPlane onClick={chatMode.chatMode === 'create' ? handleNewChatRoomCreate : handleSendMessage} />
 			</InputContainer>
 		</Container>
 	);
